// src/types/PersonSearchParams.ts
export interface PersonSearchParams {
  oip_selected: boolean | null
  oip: string | null
  enp: string | null
  ss: string | null
  doc_t: number | null
  doc_s: string | null
  doc_n: string | null
  fam: string | null
  im: string | null
  ot: string | null
  okato: string | null
  dr_from: string | null
  dr_to: string | null
  page: number | null
}